refactored e2e tests files to have 1 per crd and not many per test case per crd
