+++ test/e2e/recorder.go
@@ -26,13 +26,11 @@ var serviceGVR = schema.GroupVersionResource{Group: "", Version: "v1", Resource:
 
 func RecorderFeature(options ...eventshub.EventsHubOption) *feature.Feature {
 	to := "recorder"
-
 	f := new(feature.Feature)
 
 	a := append([]eventshub.EventsHubOption{eventshub.StartReceiver}, options...)
 	f.Setup("install recorder", eventshub.Install(to, a...))
 
 	f.Requirement("recorder is addressable", k8s.IsAddressable(serviceGVR, to, time.Second, 30*time.Second))
-
 	return f
 }
